+++ frontend/lost/src/api/pipeline/pipeline.tsx
@@ -2,9 +2,9 @@ import { useMutation, useQuery } from 'react-query'
 import { useNavigate } from 'react-router-dom'
 import { showError, showSuccess } from '../../components/Notification'
 import { httpClient } from '../http-client'
-import { StartPipeRequest } from './model/pipeline-request'
-import { PipelineResponse } from './model/pipeline-response'
-import { PipelineTemplateResponse } from './model/pipeline-template-response'
+import { StartPipeRequest } from '../../types/pipelines/pipeline-request'
+import { PipelineResponse } from '../../types/pipelines/pipeline-response'
+import { PipelineTemplateResponse } from '../../types/pipelines/pipeline-template-response'
 import {
   parseLiveElementsToReactFlow,
   parseTemplateElementsToReactFlow,
