+++ frontend/lost/src/containers/DataSources/DSTable.jsx
@@ -10,7 +10,6 @@ import BaseModal from '../../components/BaseModal'
 import LostFileBrowser from '../../components/FileBrowser/LostFileBrowser'
 import * as Notification from '../../components/Notification'
 import EditDSModal from './EditDSModal'
-// import { deleteFs } from '../../access/fb'
 import { CBadge } from '@coreui/react'
 import * as fbAPI from '../../api/fb'
 import BaseContainer from '../../components/BaseContainer'
