+++ frontend/lost/src/containers/pipeline/TourStartTable.jsx
@@ -2,7 +2,7 @@ import { CTooltip, CBadge, CRow, CCol } from '@coreui/react'
 import { useLocation } from 'react-router-dom'
 import { CContainer } from '@coreui/react'
 import { faPlay } from '@fortawesome/free-solid-svg-icons'
-import { useTemplates } from '../../actions/pipeline/pipeline_api'
+import { useTemplates } from '../../actions/pipeline/pipeline'
 import BaseContainer from '../../components/BaseContainer'
 import { CenteredSpinner } from '../../components/CenteredSpinner'
 import { useEffect, useState } from 'react'
