+++ frontend/lost/src/containers/pipeline/running/PipelineLogModal.tsx
@@ -1,4 +1,4 @@
-import { usePipelineLogs } from '../../../actions/pipeline/pipeline_api'
+import { usePipelineLogs } from '../../../actions/pipeline/pipeline'
 import { CFormTextarea } from '@coreui/react'
 import BaseModal from '../../../components/BaseModal'
