+++ frontend/lost/src/containers/pipeline/running/PipelineView.tsx
@@ -3,8 +3,8 @@ import { Node, ReactFlowProvider } from '@xyflow/react'
 import { useState } from 'react'
 import { useParams } from 'react-router-dom'
 import { useToggle } from 'react-use'
-import { PipelineResponseElement } from '../../../actions/pipeline/model/pipeline-response'
-import { usePipeline } from '../../../actions/pipeline/pipeline_api'
+import { PipelineResponseElement } from '../../../types/pipelines/pipeline-response'
+import { usePipeline } from '../../../actions/pipeline/pipeline'
 import BaseContainer from '../../../components/BaseContainer'
 import { CenteredSpinner } from '../../../components/CenteredSpinner'
 import { LivePipeline } from './LivePipeline'
