+++ frontend/lost/src/containers/pipeline/start/TemplateView.tsx
@@ -4,8 +4,8 @@ import { useState } from 'react'
 import { FaInfoCircle } from 'react-icons/fa'
 import { useNavigate, useParams } from 'react-router-dom'
 import { useToggle } from 'react-use'
-import { PipelineTemplateElement } from '../../../actions/pipeline/model/pipeline-template-response'
-import { useTemplate } from '../../../actions/pipeline/pipeline_api'
+import { PipelineTemplateElement } from '../../../types/pipelines/pipeline-template-response'
+import { useTemplate } from '../../../actions/pipeline/pipeline'
 import BaseContainer from '../../../components/BaseContainer'
 import { CenteredSpinner } from '../../../components/CenteredSpinner'
 import { NavRibbon } from './NavRibbon'
