+++ frontend/lost/src/containers/pipeline/start/modals/anno-task-modal/steps/SelectUser.tsx
@@ -1,7 +1,7 @@
 import { CBadge, CCol, CRow } from '@coreui/react'
 import { useNodesData, useReactFlow } from '@xyflow/react'
 import Select from 'react-select'
-import { AvailableGroup } from '../../../../../../actions/pipeline/model/pipeline-template-response'
+import { AvailableGroup } from '../../../../../../types/pipelines/pipeline-template-response'
 import { AnnoTaskNodeData } from '../../../nodes'
 import InfoText from '../../../../../../components/InfoText'
